requirements.txt
asgiref==3.2.10
attrs==19.3.0
autobahn==20.6.1
aws-sam-translator==1.25.0
aws-xray-sdk==2.6.0
boto==2.49.0
boto3==1.14.17
botocore==1.17.17
certifi==2020.4.5.1
cffi==1.14.0
cfn-lint==0.33.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
decorator==4.4.2
dj-database-url==0.5.0
Django==3.0.6
django-countries==6.1.2
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-mathfilters==1.0.0
django-online-users==0.3
django-social-share==1.4.0
django-storages==1.9.1
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
dnspython==1.16.0
docker==4.2.2
docutils==0.15.2
ecdsa==0.15
future==0.18.2
gunicorn==20.0.4
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.6.1
incremental==17.5.0
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
jsondiff==1.1.2
jsonpatch==1.26
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
MarkupSafe==1.1.1
mock==4.0.2
monotonic==1.5
moto==1.3.14
msgpack==0.6.2
mysql-connector-python==8.0.19
mysqlclient==1.4.6
networkx==2.4
Pillow==7.1.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jose==3.1.0
PyYAML==5.3.1
ratelim==0.1.6
reportlab==3.5.42
requests==2.23.0
responses==0.10.15
rsa==4.6
s3transfer==0.3.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
sshpubkeys==3.1.0
Twisted==20.3.0
txaio==20.4.1
urllib3==1.25.9
vine==1.3.0
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
xhtml2pdf==0.2.4
xmltodict==0.12.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
